Refactor the DeltaVocab so as to make it easier for the Lemmatizer to use it
